Refactor:split a component
