+++ src/components/idea-detail-page/Review.tsx
@@ -5,34 +5,21 @@ import { useInfiniteQuery, useMutation, useQuery, useQueryClient } from '@tansta
 import { useInView } from 'react-intersection-observer';
 
 import { getLoginUser } from '@/app/api/auth';
-import { deleteChallengeIdeaComment, getIdeaCommentInfinite, postChallengeIdeaComment, updateChallengeIdeaComment } from '@/app/api/idea-comments';
+import { getIdeaCommentInfinite, postChallengeIdeaComment } from '@/app/api/idea-comments';
 
-import DropDownBtn from './DropDownBtn';
-import { Button, useDialog } from '../common';
+import ReviewItem from './ReviewItem';
+import { Button } from '../common';
 import { Input } from '../common/Input';
 
 import type { DetailProps } from '@/app/idea/[slug]/page';
+import type { IdeaComments } from '@/types/db.type';
 
 function Review({ slug }: DetailProps) {
-  const defaultProfileImg = '../../../defaultProfileImage.jpeg';
-  const { Confirm } = useDialog();
   const [comment, setComment] = useState<string>('');
   const [userId, setUserId] = useState<string>('');
-  const [editCommentId, setEditCommentId] = useState<string>('');
-  const [editComment, setEditComment] = useState<string>('');
 
   const queryClient = useQueryClient();
   const postMutation = useMutation(postChallengeIdeaComment, {
-    onSuccess: () => {
-      queryClient.invalidateQueries(['ideaComments']);
-    },
-  });
-  const editMutation = useMutation(updateChallengeIdeaComment, {
-    onSuccess: () => {
-      queryClient.invalidateQueries(['ideaComments']);
-    },
-  });
-  const deleteMutation = useMutation(deleteChallengeIdeaComment, {
     onSuccess: () => {
       queryClient.invalidateQueries(['ideaComments']);
     },
@@ -83,24 +70,6 @@ function Review({ slug }: DetailProps) {
     setComment('');
   };
 
-  // 댓글 update
-  const handleUpdateChallengeIdeaCommentData = (id: string) => {
-    const newUpdateComment = { id, editComment };
-    editMutation.mutate(newUpdateComment);
-    setEditCommentId('');
-  };
-
-  // 댓글 delete
-  const handleDeleteChallengeIdeaCommentData = async (id: string) => {
-    const confirmed = await Confirm('해당 댓글을 삭제하시겠습니까?');
-    if (confirmed) deleteMutation.mutate(id);
-  };
-
-  const handleUpdateCommentDropDown = (id: string, comment: string) => {
-    setEditCommentId(id);
-    setEditComment(comment);
-  };
-
   if (userLoading) {
     return <p>로딩중입니다.</p>;
   }
@@ -112,39 +81,8 @@ function Review({ slug }: DetailProps) {
       <h4 className="mb-3">댓글 {challengeCommentsData?.length}</h4>
       <div className="max-h-[345px] overflow-auto">
         {challengeCommentsData?.map(commentData => {
-          const { id, created_at, comment, users } = commentData;
-          return (
-            <div key={id} className="flex items-center flex-row justify-start my-3 relative">
-              <img src={users.profile_img ? users.profile_img : defaultProfileImg} width={55} height={55} alt="Profile Image" className="mr-[16px] shadow-[0_1px_5px_0_rgba(53,60,73,0.08)] rounded-lg " />
-              <div>
-                <div className="flex flex-row text-sm text-[#838384] leading-[150%] items-center">
-                  <p className="after:content-[' '] after:w-[1px] after:bg-[#838384] after:h-[12px] after:inline-block after:m-[8px] flex items-center">{users.nickname}</p>
-                  <p>{created_at.slice(0, 10).replaceAll('-', '.')}</p>
-                </div>
-                {editCommentId === id ? (
-                  <form
-                    className="relative"
-                    onSubmit={e => {
-                      e.preventDefault();
-                    }}
-                  >
-                    <div className="absolute top-[-28px] right-0 flex">
-                      <button className="text-sm text-[#838384] after:content-[' '] after:w-[1px] after:bg-[#838384] after:h-[12px] after:inline-block after:m-[8px] flex items-center" onClick={() => handleUpdateChallengeIdeaCommentData(id)}>
-                        수정완료
-                      </button>
-                      <button className="text-sm text-[#838384]" onClick={() => setEditCommentId('')}>
-                        취소
-                      </button>
-                    </div>
-                    <Input _size="lg" type="text" value={editComment} onChange={e => setEditComment(e.target.value)} />
-                  </form>
-                ) : (
-                  <p className="leading-[150%]">{comment}</p>
-                )}
-              </div>
-              {userId === users.user_id ? <DropDownBtn editClickHandler={() => handleUpdateCommentDropDown(id, comment)} deleteClickHandler={() => handleDeleteChallengeIdeaCommentData(id)} position={'top-0 right-0'} /> : <></>}
-            </div>
-          );
+          const { id, created_at, comment, users }: IdeaComments = commentData;
+          return <ReviewItem key={id} id={id} created_at={created_at} comment={comment} users={users} user_id={userId} />;
         })}
         {hasNextPage && (
           <p className="h-[55px] flex justify-center items-center" ref={ref}>
